Separate files for abstract components
