+++ kotlin-react/src/main/kotlin/react/React.kt
@@ -1,8 +1,7 @@
 package react
 
-import kotlinext.js.*
-import kotlin.js.*
-import kotlin.reflect.*
+import kotlinext.js.asJsObject
+import kotlinext.js.jsObject
 
 external interface Child
 
@@ -33,28 +32,7 @@ inline fun <P : RProps> cloneElement(
 fun clone(element: dynamic, props: dynamic, child: Any? = null): ReactElement =
     cloneElement(element, props, *Children.toArray(child))
 
-val <P : RProps> KClass<out Component<P, *>>.rClass: RClass<P>
-    get() = js.unsafeCast<RClass<P>>()
-
 // 16.6+
 fun SuspenseProps.fallback(handler: RRender) {
     asDynamic().fallback = buildElements(handler)
 }
-
-/**
- * Usage:
- *
- * ```
- * companion object : RStatics<RProps, RState, RComponent, Nothing>(RComponent::class) {
- *     init {
- *         defaultProps = ...
- *         ...
- *     }
- * }
- * ```
- *
- * in your class components
- */
-open class RStatics<P : RProps, S : RState, C : Component<P, S>, CTX : RContext<*>?>(
-    klazz: KClass<C>
-) : RComponentClassStatics<P, S, CTX> by klazz.js.unsafeCast<RComponentClassStatics<P, S, CTX>>()
