+++ kotlin-react/src/main/kotlin/react/ReactComponent.kt
@@ -1,7 +1,6 @@
 package react
 
 import kotlinext.js.assign
-import kotlinext.js.jsObject
 
 // Props
 external interface RProps
@@ -38,20 +37,6 @@ external interface RErrorInfo
 val RErrorInfo.componentStack: Any
     get() = asDynamic().componentStack
 
-typealias RClass<P> = ComponentClass<P>
-
-external interface RComponentClassStatics<P : RProps, S : RState, C : RContext<*>?> {
-    var displayName: String?
-
-    var defaultProps: P?
-
-    var contextType: C
-
-    var getDerivedStateFromProps: ((P, S) -> S?)?
-
-    var getDerivedStateFromError: ((Throwable) -> S?)?
-}
-
 external interface RProviderProps<T> : RProps {
     var value: T
 }
@@ -80,56 +65,3 @@ external interface RReadableRef<out T : Any> : RRef {
 fun <S : RState> Component<*, S>.setState(buildState: S.() -> Unit) =
     setState({ assign(it, buildState) })
 
-abstract class RComponent<P : RProps, S : RState> : Component<P, S> {
-    constructor() : super() {
-        state = jsObject { init() }
-    }
-
-    constructor(props: P) : super(props) {
-        state = jsObject { init(props) }
-    }
-
-    open fun S.init() {}
-
-    // if you use this method, don't forget to pass props to the constructor first
-    open fun S.init(props: P) {}
-
-    fun RBuilder.children() {
-        props.children()
-    }
-
-    fun <T> RBuilder.children(value: T) {
-        props.children(value)
-    }
-
-    abstract fun RBuilder.render()
-
-    override fun render() = buildElements { render() }
-}
-
-abstract class RPureComponent<P : RProps, S : RState> : PureComponent<P, S> {
-    constructor() : super() {
-        state = jsObject { init() }
-    }
-
-    constructor(props: P) : super(props) {
-        state = jsObject { init(props) }
-    }
-
-    open fun S.init() {}
-
-    // if you use this method, don't forget to pass props to the constructor first
-    open fun S.init(props: P) {}
-
-    fun RBuilder.children() {
-        props.children()
-    }
-
-    fun <T> RBuilder.children(value: T) {
-        props.children(value)
-    }
-
-    abstract fun RBuilder.render()
-
-    override fun render() = buildElements { render() }
-}
